refactor: reuse same analysis during integration test
