+++ tests/common/helpers.py
@@ -1,4 +1,6 @@
+import random
 import time
+import uuid
 from typing import Callable
 
 from tests.common.env import PYTEST_ASYNC_MAX_RETRIES
@@ -14,3 +16,25 @@ def eventually(predicate: Callable[[], bool]) -> bool:
         return True
 
     return False
+
+
+def next_uuid():
+    return str(uuid.uuid4())
+
+
+def next_prefixed_name():
+    return f"node-it-{next_uuid()}"
+
+
+def is_valid_uuid(val: str):
+    try:
+        uuid.UUID(val)
+    except ValueError:
+        return False
+
+    return True
+
+
+def next_random_bytes(rng: random.Random, n: int = 16):
+    """Return a bytes object with random content. (default length: 16)"""
+    return rng.randbytes(n)
