+++ tests/test_upload.py
@@ -2,31 +2,10 @@ import pytest
 from starlette import status
 
 from tests.common.auth import issue_client_access_token, BearerAuth
-from tests.common.helpers import eventually
-from tests.common.rest import next_random_bytes, wrap_bytes_for_request
-from tests.test_hub import _next_prefixed_name
+from tests.common.helpers import eventually, next_random_bytes
+from tests.common.rest import wrap_bytes_for_request
 
 pytestmark = pytest.mark.live
-
-
-@pytest.fixture(scope="module")
-def analysis_id(api, rng):
-    project_name = _next_prefixed_name()
-    project = api.create_project(project_name)
-
-    analysis_name = _next_prefixed_name()
-    analysis = api.create_analysis(analysis_name, project.id)
-
-    def __check_result_bucket():
-        bucket_name = f"analysis-result-files.{analysis.id}"
-        bucket = api.get_bucket(bucket_name)
-
-        return bucket is not None
-
-    # make sure result bucket exists before firing off requests
-    assert eventually(__check_result_bucket)
-
-    yield analysis.id
 
 
 def test_200_submit_to_upload(test_client, rng, api, analysis_id):
